--- done/task2-1.rb
@@ -1,3 +1,0 @@
-File.open('movies.txt') do |file|
-  puts file.readline until file.eof?
-end
