+++ lib/cinema/cashbox.rb
