+++ lib/cinema/collection_by_country.rb
