+++ lib/cinema/collection_by_genre.rb
