+++ lib/cinema/custom_schedule.rb
