+++ lib/cinema/movie.rb
