+++ lib/cinema/movie_scraper.rb
