+++ lib/cinema/netflix.rb
