+++ lib/cinema/period.rb
