+++ lib/cinema/schedule.rb
