+++ lib/cinema/schedule_builder.rb
