+++ lib/cinema/templates/movie_collection.html.erb
