+++ lib/cinema/theatre.rb
